PHPFront now implements the JQuery API
On the quest for a simpler life with code... We're here now! PHPFront
now implements the JQuery API. That means that anything you know how to
do with JQuery on the front end is now also possible on the backend. If
you can do $("title").html("This is page title"), then you can also say
$phpFront->find('title')->html('This is page title').
This is a total rewrite of phpFront! No backwards compatibility. But
this is certainly the future of server-side rendering!